refactor(workflows): move slot and phase-label orchestration out of lib/
Place GitHub-facing helpers under issue-slots/compute.js and phase-label/set.js;
keep pure logic in lib/issue-slots.js and lib/phase-label.js.
